pyproject.toml
[tool.poetry]
name = "mhw_booklet"
version = "1.0.0"
description = "Booklet that shows weaknesses/attack types of bosses"
authors = ["Алексей Чистов <aleksey.chistov@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
Pillow = "^8.2.0"
commentjson = "^0.9.0"
structlog = "^21.1.0"
colorama = "^0.4.4"
progress = "^1.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
wemake-python-styleguide = "^0.15.2"

rope = "^0.19.0"
isort = ">=5"
unify = "^0.5"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
